+++ webpack.mix.js
@@ -6,6 +6,7 @@ mix.setPublicPath('./public/default/');
 mix.sass('resources/assets/scss/app.scss', 'css/admin-app.css')
   .js('resources/assets/js_owl/vue_init.js', 'js/vue.js')
   .js('resources/assets/js_owl/app.js', 'js/admin-app.js')
+  .js('resources/assets/js_owl/app-dev.js', 'js/admin-app-dev.js')
   .js('resources/assets/js_owl/modules_load.js', 'js/modules.js')
   .options({
     processCssUrls: true,
@@ -18,6 +19,4 @@ mix.sass('resources/assets/scss/app.scss', 'css/admin-app.css')
 
 if (mix.inProduction()) {
   mix.version()
-} else {
-  mix.copy('public/default', '../../../public/packages/sleepingowl/default')
 }