+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # [Unreleased] (Only in SleepingOwl <code class="language-php">8+</code> branch)
 
 ## 2021-11-15
+* [Add] Add class for `icheck indeterminate`
+* [Rebuild] Rebuild assets and add use `dev_assets` in config (use `ADMIN_DEV_ASSETS=true` in `.env`)
+
+## 2021-11-15
 * [Add] Add `->setTarget(string)` for Navigations
 
 ## 2021-10-03